+++ src/pages/Parameter.vue
@@ -38,7 +38,7 @@
               color="grey"
               size="sm"
               dense round icon="undo"
-              @click="refreshParameter"
+              @click="undoParameter"
             />
             <q-btn
               class="q-ml-md"
@@ -81,10 +81,10 @@
                       ref="parameter-value"
                       class="q-pa-xs"
                       style="font-family:monospace;min-width:24em;word-break:break-all"
-                      :param-name="parameter.paramName"
+                      :param-name="params.paramName"
                       :param-type="parameter.type"
                       :param-value="parameter.defaultValue"
-                      :default-value="original.defaultValue"
+                      :original-value="original.defaultValue"
                       editable
                       @save="val => { parameter.defaultValue = val }"
                       @cancel="() => { parameter.defaultValue = original.defaultValue }"
@@ -208,6 +208,10 @@
     },
 
     methods: {
+      undoParameter() {
+        this.parameter = cloneDeep(this.original)
+      },
+
       refreshParameter() {
         this.$apollo.queries.parameter.refetch()
       },